requirements.txt
# Core
numpy>=1.23
scipy
tqdm
pyyaml
einops

# Image + data
pillow
opencv-python
scikit-image

# ML
torch>=2.1
torchvision
torchaudio

# Logging / analysis
matplotlib
tensorboard

# Utilities
rich
psutil
